Accommodate mixed feedback for same character
Consider the following simulated/redacted game I observed:
```
_ _ a _ _
_ _ . _ _
_ a _ a _
_ = _ - _
```

Once a letter is placed, a misplaced instance of the same letter will be
reported as not appearing in the word, as long as the letter doesn't
appear twice in the word.

This general pattern is modeled as two new types of facts:

- Exclude where not placed: a placed letter is known not to appear
anywhere else in the word

- Minimum occurrence count: Mixed placed and misplaced feedback for the
same character indicates the letter appears at least that many times.
This is a fact that can only be generated by a whole turn, not local
feedback information.